crawler/requirements.txt
Scrapy==1.7.4
selenium==3.141.0
google-cloud-pubsub==0.45.0